Perf -> Master: Engine restructuring, GPU-based lighting, and build system overhaul

+++ src/ECS/Components/DirectionalTextureComponent.h
@@ -9,7 +9,6 @@ namespace ECS::Components {
     // for 6-way rotation sprite sheet stuff
     struct DirectionalTextureComponent {
         std::array<std::shared_ptr<Rendering::Texture>, 6> textures;
-        int index = 0;
+        uint8_t index = 0;
     };
 } // namespace ECS::Components
-
